clean main
